Added test for invalid port and minor code cleanup
